requirements.txt
chromadb
sentence-transformers
pypdf
python-docx
scikit-learn